requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
pytest
httpx
sqlalchemy
python-dotenv